refactor: Move loss and metric to trainer
- move loss and metric to traniner folder
- remove arch folder and move models to model folder

#35
