--- model/arch/__init__.py
@@ -1,2 +1,0 @@
-from .torchvision_model import *
-from .smp_model import *
