+++ model/torchvision_model.py
