+++ trainer/__init__.py
@@ -1,1 +1,3 @@
 from .trainer import *
+from .loss import *
+from .metric import *
